goodtxt/backend/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# AI and LLM integrations
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# Database and storage
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
chromadb>=0.4.0

# Authentication and security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
python-multipart>=0.0.6

# Configuration and environment
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Data processing and utilities
numpy>=1.24.0
scipy>=1.11.0

# Logging and monitoring
structlog>=23.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0